add(chunk): persist real per-copy width instead of hardcoding it
Every backend's own per-copy metadata (file::Backend's .meta file,
lvm::Backend's LV tag, zfs::Backend's ZFS property) used to hardcode
width=1 in its own encoded record regardless of the object's real
width; blk::Backend::meta_encode()/meta_decode() now round-trip a
ChunkIdentity{width} out-parameter instead, stamped once at create()
and never changed again, so each copy's record actually reflects the
target's own shape. RawstorOSTFrameAllocatePayload/MetaPayload gain a
matching `width` field over the wire, so rawstor-ost's own copies get
the same treatment.

This has no observable effect yet: Target::spec()/meta() always
overwrite spec.width with the group's own URI count rather than trust
any single copy's report, since width isn't a property of one copy.
It's still worth doing now rather than leaving every backend to
special-case width=1 -- the plumbing (ChunkIdentity, meta_encode()/
meta_decode(), the wire fields) is the same regardless of how many
real fields it carries.

Chunk placement identity itself (chunk_size, member_kind -- docs/
mds.md, chunk_meta) deliberately stays out for now: nothing here or in
any bucket built so far has a use for either, so both are left for
whichever later bucket actually introduces a real consumer, rather
than carried as dead API surface until then. The wire structs reserve
generic space for them instead (reserved1..4 on Allocate, reserved1..3
on Meta), so they can be added back without another layout change.
META_MAX_SIZE stays 256 (comfortably covers the current field set);
META_FORMAT_VERSION stays 1 -- this whole per-copy metadata format is
itself still part of the unreleased 0.3.0 line, nothing to stay
compatible with yet.

Also deliberately excluded, left for add/mds-protocol-ported itself:
stripe_width/failure_domain (placement policy, not identity), the CLI
--chunk-size/--failure-domain/--stripe-width flags, the MDS server,
the OBJ_* wire commands, and any multi-chunk routing.

+++ tests/session.cpp
@@ -282,7 +282,7 @@ void Session::cmd_spec_response(
 
     RawstorOSTFrameSpecPayload spec = {
         .size = size,
-        .width = static_cast<uint32_t>(width),
+        .width = static_cast<uint8_t>(width),
     };
     RawstorOSTFrameResponse response = {
         .head{
